Rename package cli to cmd
Seems to be the default name when using cobra.
